requirements.txt
Tornado==5.0.2
PySide2==5.11
pyserial==3.4